maskRCNN-resnet-50-fpn/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.18.0
matplotlib>=3.3.0
pillow>=8.0.0
opencv-python>=4.5.0
tqdm>=4.50.0
tensorboard>=2.4.0
pycocotools>=2.0.2
scikit-image>=0.17.0